Load JS after DOM is complete
Load only the bare minimum JS to enable the pages.html fragment
to be loaded; once that HTML is loaded and in the DOM,
load the majority of the JS files; once they're all done,
init the app.

This provides a substantial benefit, in that the initial display
of the HTML page is about 3 times faster.